requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
redis
langchain-redis